Add `patch_config` function with unit test

+++ test/test_configuration.py
@@ -58,7 +58,8 @@ class ConfigurationTests(unittest.TestCase):
 
     @classmethod
     def tearDownClass(cls) -> None:
-        shutil.rmtree(cls.CONFIG_PATH)
+        if os.path.exists(cls.CONFIG_PATH):
+            shutil.rmtree(cls.CONFIG_PATH)
         os.environ.pop("XDG_CONFIG_HOME")
 
     def test_neon_core_config_init(self):
@@ -107,6 +108,38 @@ class ConfigurationTests(unittest.TestCase):
         self.assertEqual(neon_ipc_dir, use_neon_core(ovos_ipc_dir)())
         self.assertEqual(neon_ipc_dir,
                          use_neon_core(mycroft_ipc_dir)())
+
+    def test_patch_config(self):
+        from os.path import join
+        import json
+
+        test_config_dir = os.path.join(os.path.dirname(__file__), "config")
+        os.makedirs(test_config_dir, exist_ok=True)
+        os.environ["XDG_CONFIG_HOME"] = test_config_dir
+
+        from neon_core.util.runtime_utils import use_neon_core
+        from neon_utils.configuration_utils import init_config_dir
+
+        use_neon_core(init_config_dir)()
+
+        with open(join(test_config_dir, "OpenVoiceOS", 'ovos.conf')) as f:
+            ovos_conf = json.load(f)
+        self.assertEqual(ovos_conf['submodule_mappings']['neon_core'],
+                         "neon_core")
+        self.assertIsInstance(ovos_conf['module_overrides']['neon_core'], dict)
+
+        from neon_core.configuration import patch_config
+        test_config = {"new_key": {'val': True}}
+        patch_config(test_config)
+        conf_file = os.path.join(test_config_dir, 'neon',
+                                 'neon.conf')
+        self.assertTrue(os.path.isfile(conf_file))
+        with open(conf_file) as f:
+            config = json.load(f)
+
+        self.assertTrue(config['new_key']['val'])
+        shutil.rmtree(test_config_dir)
+        # os.environ.pop("XDG_CONFIG_HOME")
 
 
 if __name__ == '__main__':
